requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
Django==3.1.1
djangorestframework==3.11.1
isort==5.5.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
django-allauth==0.42.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
